refactor: 修正 XivMaps 的导出方式以解决 `Cannot access before initialization` 错误 * 但是为啥现在才报错？尖尖瓦甘来
