+++ src/components/custom/map/LocationSpan.vue
@@ -1,7 +1,7 @@
 <script setup lang="ts">
 import { computed, inject, type Ref } from 'vue'
 import MapButton from './MapButton.vue'
-import { XivMaps } from '@/assets/data'
+import { XivMaps } from '@/tools/map'
 import type { UserConfigModel } from '@/models/config-user'
 
 const userConfig = inject<Ref<UserConfigModel>>('userConfig')!
