+++ src/views/GatherClockPage.vue
@@ -15,10 +15,11 @@ import ItemButton from '@/components/custom/item/ItemButton.vue'
 import XivMap from '@/components/custom/map/XivMap.vue'
 import LocationSpan from '@/components/custom/map/LocationSpan.vue'
 import ModalAlarmMacroExport from '@/components/modals/ModalAlarmMacroExport.vue'
-import { XivMaps, XivJobs, type XivJob } from '@/assets/data'
+import { XivJobs, type XivJob } from '@/assets/data'
 import { useStore } from '@/store'
 import { useNbbCal } from '@/tools/use-nbb-cal'
 import { getItemInfo, type ItemInfo } from '@/tools/item'
+import { XivMaps } from '@/tools/map'
 import useUiTools from '@/tools/ui'
 import type { ItemGroup } from '@/models/item'
 import type { UserConfigModel } from '@/models/config-user'
